Week:
Implemented functions for, AverageExpected weight, AverageCheckinWeight, WeeklyCalories, A create week function, fucntions to get week and profile items from DB.

Implemented Action for Week1 and Week 2

Next is to implement action and view for Week 2 and Week 3.

Models:

Updated models to be able to store all the required values, migrated and updated database. The database is now up to date.